Moved main resolve logic to protocol extensions instead of commands
