+++ server/src/main/kotlin/org/javacs/kt/command/Commands.kt
@@ -1,9 +1,7 @@
 package org.javacs.kt.command
 
 const val JAVA_TO_KOTLIN_COMMAND = "convertJavaToKotlin"
-const val RESOLVE_MAIN = "resolveMain"
 
 val ALL_COMMANDS = listOf(
     JAVA_TO_KOTLIN_COMMAND,
-    RESOLVE_MAIN
 )
